Updated some files
Worked on RC.c and Timer.c

+++ Paperino_3.1/Paperino_3.1/Timer.h
@@ -45,9 +45,9 @@ void timer_8bit_INT_init_OVF();
 
 void timer_8bit_CTC_init(uint8_t freq, uint16_t pres);
 
-uint16_t time_in_milliseconds(uint16_t precision, uint16_t precision_module);
-uint16_t time_in_seconds(uint16_t precision, uint16_t precision_module);
-uint16_t time_precision(uint16_t precision, uint16_t precision_module);
+uint32_t time_in_milliseconds(uint16_t precision, uint16_t precision_module);
+uint32_t time_in_seconds(uint16_t precision, uint16_t precision_module);
+uint32_t time_precision(uint16_t precision, uint16_t precision_module);
 
 extern volatile uint16_t system_tick_MG;
 extern volatile uint16_t system_tick_MG_p;
